make resources a submodule
