+++ collada/schema.py
@@ -14,36 +14,11 @@
 """This module contains helper classes and functions for working
 with the COLLADA 1.4.1 schema."""
 
-import sys
 import lxml
 import lxml.etree
 from collada.util import bytes, BytesIO
 
-
-def resource_string(file_name: str) -> str:
-    """
-    Get the value of a file in `collada/resources/{file_name}`
-    as a string.
-
-    Parameters
-    -----------
-    file_name
-      The name of the file in `collada/resources/{file_name}`
-
-    Returns
-    ----------
-    value
-      The contents of the file.
-    """
-    if sys.version_info <= (3, 9):
-        from importlib.resources import read_text
-
-        return read_text("collada.resources", file_name)
-    else:
-        from importlib import resources
-
-        return resources.files("collada").joinpath("resources", file_name).read_text()
-
+from .resources import resource_string
 
 # get a copy of the XML schema
 COLLADA_SCHEMA_1_4_1 = resource_string("schema-1.4.1.xml")
